test(activemodel): move the TS-only extras out of decimal, date and big-integer type tests
